refactor(routes): /api/users replaces /api/cards (#55)
Endpoints:
- GET    /api/users (list, staff)
- POST   /api/users (create, staff; replaces activate-card)
- GET    /api/users/search (staff)
- GET    /api/users/lookup/{code} (staff)
- POST   /api/users/topup (staff)
- POST   /api/users/block (staff)
- GET    /api/users/negative-balance (staff)
- PUT    /api/users/{id} (staff)
- GET    /api/users/{id}/transactions (staff or owner)
- GET    /api/users/{id}/stats (staff)
- GET    /api/my/balance (any auth)

Email is now optional at create. card_code is the legacy chip reference.
All visit subqueries bind CLASS_VISIT_NAMES_EN.
