+++ crates/spinbike-server/src/routes/mod.rs
@@ -1,6 +1,5 @@
 pub mod admin;
 pub mod auth;
-pub mod cards;
 pub mod classes;
 pub mod payments;
 pub mod persistent_bookings;
@@ -9,6 +8,7 @@ pub mod static_files;
 pub mod test_fixtures;
 pub mod transactions;
 pub mod upcoming_classes;
+pub mod users;
 pub mod version;
 
 use axum::{Json, Router, http::StatusCode};
@@ -43,7 +43,7 @@ pub fn api_routes() -> Router<AppState> {
     Router::new()
         .merge(auth::routes())
         .merge(classes::routes())
-        .merge(cards::routes())
+        .merge(users::routes())
         .merge(payments::routes())
         .merge(admin::routes())
         .merge(persistent_bookings::routes())
